Fix warnings on sorting algorithm unit tests
This commit fixes warnings present in the unit testing suites of the
bubblesort implementation (`bubblesort`) and the insertion sort
implementation (`isort`).

+++ tests/unit/bubblesort/unit_bubblesort.c
@@ -15,7 +15,7 @@
 
 bool cmp_clbk(void* a, void* b)
 {
-    if((unsigned int)a >= (unsigned int)b)
+    if((intptr_t)a >= (intptr_t)b)
     {
         return true;
     }
